feat: separate the trait of FovEngine with the implementation of SquidLibFov
